ENG-01: extract TPLS/DataFrameDict into _tpls; move Plot accessor to plots
Move the T-PLS estimator and its DataFrameDict container into _tpls, and
relocate the Plot accessor class to plots.py (its only user is TPLS; keeping
it in _resampling would have created a _tpls<->_resampling cycle since
Resampler imports DataFrameDict). Drop the dead commented-out
_calculate_r2_score block, and remove now-unused imports from _pca_pls.
